needs to be int
old-commit-hash: ff1cd3de1e380ba3990ff70c231bf99d1272a331

+++ selfdrive/hardware/tici/hardware.py
@@ -187,6 +187,6 @@ class Tici(HardwareBase):
   def set_screen_brightness(self, percentage):
     try:
       with open("/sys/class/backlight/panel0-backlight/brightness", "w") as f:
-        f.write(str(percentage * 10.23))
+        f.write(str(int(percentage * 10.23)))
     except Exception:
       pass
